abstract invite-specific content from AuthenticationView
